Major rework done
